requirements.txt
Django==3.2.12
djangorestframework==3.12.4
psycopg2-binary==2.9.2
django-cors-headers==3.7.0